Fixed some potential issues with upgrade routine.

+++ includes/admin/upgrades.php
@@ -55,18 +55,21 @@ function nss_do_upgrade( $db_version ) {
 	 * 1.2.0 - Add new social sites.
 	 */
 	if ( version_compare( $db_version, '1.2.0', '<' ) ) {
-		$old_settings = get_option( 'naked_ss__settings', array() );
-		$social_sites = ( array_key_exists( 'social_sites', $old_settings ) ) ? $old_settings['social_sites'] : array();
+		$old_settings = get_option( 'naked_ss__settings' );
 
-		// Add LinkedIn.
-		if ( ! array_key_exists( 'linkedin', $social_sites['enabled'] ) && ! array_key_exists( 'linkedin', $social_sites['disabled'] ) ) {
-			$social_sites['disabled']['linkedin'] = array(
-				'name' => __( 'LinkedIn', 'naked-social-share' )
-			);
+		if ( ! empty( $old_settings ) ) {
+			$social_sites = ( array_key_exists( 'social_sites', $old_settings ) ) ? $old_settings['social_sites'] : array();
 
-			$old_settings['social_sites'] = $social_sites;
+			// Add LinkedIn.
+			if ( is_array( $social_sites['enabled'] ) && ! array_key_exists( 'linkedin', $social_sites['enabled'] ) && ! array_key_exists( 'linkedin', $social_sites['disabled'] ) ) {
+				$social_sites['disabled']['linkedin'] = array(
+					'name' => __( 'LinkedIn', 'naked-social-share' )
+				);
 
-			update_option( 'naked_ss__settings', $old_settings );
+				$old_settings['social_sites'] = $social_sites;
+
+				update_option( 'naked_ss__settings', $old_settings );
+			}
 		}
 	}
 
@@ -76,23 +79,26 @@ function nss_do_upgrade( $db_version ) {
 	 */
 	if ( version_compare( $db_version, '1.3.0', '<' ) ) {
 		$old_settings = get_option( 'naked_ss__settings', array() );
-		$social_sites = ( array_key_exists( 'social_sites', $old_settings ) ) ? $old_settings['social_sites']['enabled'] : array();
 
-		if ( array_key_exists( 'placebo', $social_sites ) ) {
-			unset( $social_sites['placebo'] );
+		if ( ! empty( $old_settings ) ) {
+			$social_sites = ( array_key_exists( 'social_sites', $old_settings ) ) ? $old_settings['social_sites']['enabled'] : array();
+
+			if ( array_key_exists( 'placebo', $social_sites ) ) {
+				unset( $social_sites['placebo'] );
+			}
+
+			if ( array_key_exists( 'enabled', $old_settings ) ) {
+				unset( $old_settings['enabled'] );
+			}
+			if ( array_key_exists( 'disabled', $old_settings ) ) {
+				unset( $old_settings['disabled'] );
+			}
+
+			$old_settings['social_sites'] = array_keys( $social_sites );
+			$new_settings                 = $old_settings;
+
+			update_option( 'naked_social_share_settings', $new_settings );
 		}
-
-		if ( array_key_exists( 'enabled', $old_settings ) ) {
-			unset( $old_settings['enabled'] );
-		}
-		if ( array_key_exists( 'disabled', $old_settings ) ) {
-			unset( $old_settings['disabled'] );
-		}
-
-		$old_settings['social_sites'] = array_keys( $social_sites );
-		$new_settings                 = $old_settings;
-
-		update_option( 'naked_social_share_settings', $new_settings );
 	}
 
 	/*
